@@ -1,10 +1,7 @@
 name: url_launcher_web
 description: Web platform implementation of url_launcher
 homepage: https://github.com/flutter/plugins/tree/master/packages/url_launcher/url_launcher_web
-# 0.1.y+z is compatible with 1.0.0, if you land a breaking change bump
-# the version to 2.0.0.
-# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
-version: 0.1.5+3
+version: 2.0.0-nullsafety
 
 flutter:
   plugin:
@@ -14,31 +11,18 @@ flutter:
         fileName: url_launcher_web.dart
 
 dependencies:
-  url_launcher_platform_interface: ^1.0.9
-  # TODO(mvanbeusekom): Update to use pub.dev once null safety version is published.
-  # url_launcher_platform_interface:
-  #   git:
-  #     url: https://github.com/flutter/plugins.git
-  #     ref: nnbd
-  #     path: packages/url_launcher/url_launcher_platform_interface
+  url_launcher_platform_interface: ^2.0.0-nullsafety
+  meta: ^1.3.0 # null safe
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  meta: ^1.1.7
 
 dev_dependencies:
+  pedantic: ^1.10.0 # null safe
   flutter_test:
     sdk: flutter
-  url_launcher: ^5.2.5
-  # TODO(mvanbeusekom): Update to use pub.dev once null safety version is published.
-  # url_launcher:
-  #   path: ../url_launcher
-  pedantic: ^1.8.0
-  mockito: ^4.1.1
-  integration_test:
-    path: ../../integration_test
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
