@@ -2,34 +2,35 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.9
 import 'dart:html' as html;
-import 'dart:js_util';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:url_launcher_platform_interface/link.dart';
+import 'package:mockito/annotations.dart';
 import 'package:url_launcher_web/url_launcher_web.dart';
-import 'package:url_launcher_web/src/link.dart';
 import 'package:mockito/mockito.dart';
 import 'package:integration_test/integration_test.dart';
 
-class _MockWindow extends Mock implements html.Window {}
+import 'url_launcher_web_test.mocks.dart';
 
-class _MockNavigator extends Mock implements html.Navigator {}
-
+@GenerateMocks([html.Window, html.Navigator])
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   group('UrlLauncherPlugin', () {
-    _MockWindow mockWindow;
-    _MockNavigator mockNavigator;
+    late MockWindow mockWindow;
+    late MockNavigator mockNavigator;
 
-    UrlLauncherPlugin plugin;
+    late UrlLauncherPlugin plugin;
 
     setUp(() {
-      mockWindow = _MockWindow();
-      mockNavigator = _MockNavigator();
+      mockWindow = MockWindow();
+      mockNavigator = MockNavigator();
       when(mockWindow.navigator).thenReturn(mockNavigator);
+
+      // Simulate that window.open does something.
+      when(mockWindow.open(any, any)).thenReturn(MockWindow());
+
+      when(mockNavigator.vendor).thenReturn('Google LLC');
+      when(mockNavigator.appVersion).thenReturn('Mock version!');
 
       plugin = UrlLauncherPlugin(debugWindow: mockWindow);
     });
@@ -60,27 +61,10 @@ void main() {
     });
 
     group('launch', () {
-      setUp(() {
-        // Simulate that window.open does something.
-        when(mockWindow.open('https://www.google.com', ''))
-            .thenReturn(_MockWindow());
-        when(mockWindow.open('mailto:name@mydomain.com', ''))
-            .thenReturn(_MockWindow());
-        when(mockWindow.open('tel:5551234567', '')).thenReturn(_MockWindow());
-        when(mockWindow.open('sms:+19725551212?body=hello%20there', ''))
-            .thenReturn(_MockWindow());
-      });
-
       testWidgets('launching a URL returns true', (WidgetTester _) async {
         expect(
             plugin.launch(
               'https://www.google.com',
-              useSafariVC: null,
-              useWebView: null,
-              universalLinksOnly: null,
-              enableDomStorage: null,
-              enableJavaScript: null,
-              headers: null,
             ),
             completion(isTrue));
       });
@@ -89,12 +73,6 @@ void main() {
         expect(
             plugin.launch(
               'mailto:name@mydomain.com',
-              useSafariVC: null,
-              useWebView: null,
-              universalLinksOnly: null,
-              enableDomStorage: null,
-              enableJavaScript: null,
-              headers: null,
             ),
             completion(isTrue));
       });
@@ -103,12 +81,6 @@ void main() {
         expect(
             plugin.launch(
               'tel:5551234567',
-              useSafariVC: null,
-              useWebView: null,
-              universalLinksOnly: null,
-              enableDomStorage: null,
-              enableJavaScript: null,
-              headers: null,
             ),
             completion(isTrue));
       });
@@ -117,12 +89,6 @@ void main() {
         expect(
             plugin.launch(
               'sms:+19725551212?body=hello%20there',
-              useSafariVC: null,
-              useWebView: null,
-              universalLinksOnly: null,
-              enableDomStorage: null,
-              enableJavaScript: null,
-              headers: null,
             ),
             completion(isTrue));
       });
@@ -232,121 +198,5 @@ void main() {
         });
       });
     });
-  });
-
-  group('link', () {
-    testWidgets('creates anchor with correct attributes',
-        (WidgetTester tester) async {
-      final Uri uri = Uri.parse('http://foobar/example?q=1');
-      await tester.pumpWidget(Directionality(
-        textDirection: TextDirection.ltr,
-        child: WebLinkDelegate(TestLinkInfo(
-          uri: uri,
-          target: LinkTarget.blank,
-          builder: (BuildContext context, FollowLink followLink) {
-            return Container(width: 100, height: 100);
-          },
-        )),
-      ));
-      // Platform view creation happens asynchronously.
-      await tester.pumpAndSettle();
-
-      final html.Element anchor = _findSingleAnchor();
-      expect(anchor.getAttribute('href'), uri.toString());
-      expect(anchor.getAttribute('target'), '_blank');
-
-      final Uri uri2 = Uri.parse('http://foobar2/example?q=2');
-      await tester.pumpWidget(Directionality(
-        textDirection: TextDirection.ltr,
-        child: WebLinkDelegate(TestLinkInfo(
-          uri: uri2,
-          target: LinkTarget.self,
-          builder: (BuildContext context, FollowLink followLink) {
-            return Container(width: 100, height: 100);
-          },
-        )),
-      ));
-      await tester.pumpAndSettle();
-
-      // Check that the same anchor has been updated.
-      expect(anchor.getAttribute('href'), uri2.toString());
-      expect(anchor.getAttribute('target'), '_self');
-    });
-
-    testWidgets('sizes itself correctly', (WidgetTester tester) async {
-      final Key containerKey = GlobalKey();
-      final Uri uri = Uri.parse('http://foobar');
-      await tester.pumpWidget(Directionality(
-        textDirection: TextDirection.ltr,
-        child: Center(
-          child: ConstrainedBox(
-            constraints: BoxConstraints.tight(Size(100.0, 100.0)),
-            child: WebLinkDelegate(TestLinkInfo(
-              uri: uri,
-              target: LinkTarget.blank,
-              builder: (BuildContext context, FollowLink followLink) {
-                return Container(
-                  key: containerKey,
-                  child: SizedBox(width: 50.0, height: 50.0),
-                );
-              },
-            )),
-          ),
-        ),
-      ));
-      await tester.pumpAndSettle();
-
-      final Size containerSize = tester.getSize(find.byKey(containerKey));
-      // The Stack widget inserted by the `WebLinkDelegate` shouldn't loosen the
-      // constraints before passing them to the inner container. So the inner
-      // container should respect the tight constraints given by the ancestor
-      // `ConstrainedBox` widget.
-      expect(containerSize.width, 100.0);
-      expect(containerSize.height, 100.0);
-    });
-  });
-}
-
-html.Element _findSingleAnchor() {
-  final List<html.Element> foundAnchors = <html.Element>[];
-  for (final html.Element anchor in html.document.querySelectorAll('a')) {
-    if (hasProperty(anchor, linkViewIdProperty)) {
-      foundAnchors.add(anchor);
-    }
-  }
-
-  // Search inside platform views with shadow roots as well.
-  for (final html.Element platformView
-      in html.document.querySelectorAll('flt-platform-view')) {
-    final html.ShadowRoot shadowRoot = platformView.shadowRoot;
-    if (shadowRoot != null) {
-      for (final html.Element anchor in shadowRoot.querySelectorAll('a')) {
-        if (hasProperty(anchor, linkViewIdProperty)) {
-          foundAnchors.add(anchor);
-        }
-      }
-    }
-  }
-
-  return foundAnchors.single;
-}
-
-class TestLinkInfo extends LinkInfo {
-  @override
-  final LinkWidgetBuilder builder;
-
-  @override
-  final Uri uri;
-
-  @override
-  final LinkTarget target;
-
-  @override
-  bool get isDisabled => uri == null;
-
-  TestLinkInfo({
-    @required this.uri,
-    @required this.target,
-    @required this.builder,
   });
 }
